avances en Backend

+++ TBDBackendLab1/src/main/java/com/example/TBDBackendLab1/persistence/repository/CategoryRepository.java
@@ -5,5 +5,5 @@ import com.example.TBDBackendLab1.persistence.entity.CategoryEntity;
 public interface CategoryRepository {
 
     CategoryEntity addCategory(CategoryEntity category);
-    CategoryEntity getById(Long id);
+    CategoryEntity getById(Integer id);
 }
